pyproject.toml
[tool.poetry]
name = "streemer"
version = "0.1.0"
description = "Utilities to query streem api"
authors = ["Sébastien Julliot <sjulliot@streem.eu>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
simplejson = "^3.19.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
